初版本已完成 Signed-off-by: xiaochenHomePc <378540108@qq.com>

+++ main/nvs_app.h
@@ -15,13 +15,13 @@ typedef struct _parameter   //掉电不丢失参数，//对应mbdata0-64
     uint16_t mode;      //工作模式 低四位  0：232  1：485 2：wifi 3:BLE 
     uint16_t station;   //站号   
     uint16_t coefficient;   //adc转重量系数
-    uint16_t zero_error;  //零误差
-    uint16_t skin;        //皮重
+    int32_t zero_error;  //零误差
+    uint32_t skin;        //皮重
 
 }Parameter;
 typedef struct _realtime_value  //对应mbdata63-80
 {
-    uint16_t weight;      //重量 单位g
+    int32_t weight;      //重量 单位g
     uint16_t coefficient; //置零 执行完后会自动复位
     uint16_t remove_skin; //去皮 执行完后会自动复位  
     uint16_t calibrate_start; //开始校准 当被置1后 程序会记录当前adc值
